A type of 'id' field has been replaced from the 'UserStatus' class. The 'getMap' method has been added into the 'UserStatus' interface. It has been implemented into the implementing classes

+++ src/main/java/dao/impl/mysql/UserDaoImpl.java
@@ -96,7 +96,7 @@ class UserDaoImpl extends AbstractDao<User> implements UserDao {
         ps.setString(2, entity.getName());
         ps.setString(3, entity.getSurname());
         ps.setString(4, entity.getPassword());
-        ps.setInt(5, entity.getStatus().getId());
+        ps.setLong(5, entity.getStatus().getId());
 
         return ps;
